Rearranged functions into newly created workflow.py, remove some code duplication.
